Move EC2 customizations into a single directory
