+++ awscli/customizations/ec2/protocolarg.py
@@ -16,6 +16,7 @@ or "icmp" as values for the --protocol parameter.  The actual Protocol
 parameter of the operation accepts only integer protocol numbers.
 """
 
+
 def _fix_args(params, **kwargs):
     key_name = 'Protocol'
     if key_name in params:
